requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.10.0
itsdangerous==2.2.0
marshmallow-sqlalchemy==1.1.0
marshmallow==3.23.1
mysqlclient==2.2.6
python-dotenv==1.0.1
Werkzeug==3.1.3